# Conflicts:
#	event-service/src/main/java/com/ed/eventservice/coupon/adapter/out/persistence/entity/CouponDiscountInfoJpaEntity.java
#	event-service/src/main/java/com/ed/eventservice/coupon/adapter/out/persistence/entity/CouponExpirationInfoJpaEntity.java
#	event-service/src/main/java/com/ed/eventservice/coupon/adapter/out/persistence/entity/CouponIssueInfoJpaEntity.java
#	event-service/src/main/java/com/ed/eventservice/coupon/adapter/out/persistence/entity/CouponUsageTargetInfoJpaEntity.java
#	payment-service/build.gradle
#	payment-service/src/main/java/com/ed/payment/application/port/in/ConfirmPaymentCommand.java
#	payment-service/src/main/java/com/ed/payment/application/port/out/persistence/CreatePaymentHistoryPort.java
#	payment-service/src/main/java/com/ed/payment/application/port/out/persistence/CreatePaymentPort.java
#	payment-service/src/main/java/com/ed/payment/application/port/out/persistence/ReadPaymentPort.java
#	payment-service/src/main/java/com/ed/payment/application/port/out/persistence/UpdatePaymentPort.java
#	payment-service/src/main/java/com/ed/payment/application/port/out/pg/PaymentDone.java
#	payment-service/src/main/java/com/ed/payment/application/service/ConfirmPaymentService.java
#	payment-service/src/main/java/com/ed/payment/application/service/HandlerFailPaymentService.java
#	payment-service/src/main/java/com/ed/payment/domain/Payment.java
#	payment-service/src/main/java/com/ed/payment/infrastructure/out/pg/toss/RestTemplateTossConfirmPayment.java
#	payment-service/src/main/java/com/ed/payment/libs/common/exception/ErrorCode.java
#	payment-service/src/main/java/com/ed/payment/libs/common/response/ApiResponseUtils.java
#	payment-service/src/main/java/com/ed/payment/libs/common/response/CommonHttpMessageConverter.java
#	payment-service/src/main/java/com/ed/payment/libs/common/response/ErrorResponse.java
#	payment-service/src/main/java/com/ed/payment/presentation/in/web/ConfirmPaymentController.java
#	payment-service/src/main/java/com/ed/payment/presentation/in/web/HandleFailPaymentController.java
#	payment-service/src/main/resources/application-local.yml
#	payment-service/src/main/resources/http/payment.http
#	payment-service/src/main/resources/templates/checkout.html
#	payment-service/src/test/java/com/ed/payment/application/service/ConfirmPaymentServiceTest.java
#	payment-service/src/test/java/com/ed/payment/application/service/HandlerFailPaymentServiceTest.java
#	payment-service/src/test/java/com/ed/payment/infrastructure/out/pg/toss/RestTemplateTossConfirmPaymentTest.java
#	payment-service/src/test/java/com/ed/payment/presentation/in/web/ConfirmPaymentControllerTest.java
#	payment-service/src/test/java/com/ed/payment/presentation/in/web/HandleFailPaymentControllerTest.java

+++ payment-service/src/main/java/com/ed/payment/application/port/in/ConfirmPaymentCommand.java
@@ -14,10 +14,10 @@ public class ConfirmPaymentCommand {
   private String paymentType;
   private String paymentKey;
   private String orderId;
-  private int amount;
+  private Long amount;
 
   public static ConfirmPaymentCommand of(
-      String paymentType, String paymentKey, String orderId, int amount) {
+      String paymentType, String paymentKey, String orderId, Long amount) {
     return ConfirmPaymentCommand.builder()
         .paymentType(paymentType)
         .paymentKey(paymentKey)
